fix: separate card button
