+++ src/components/shared/card-item/card-button/index.js
@@ -1,0 +1,3 @@
+import CardButton from './card-button';
+
+export default CardButton;
