+++ src/components/shared/card-item/card-item.jsx
@@ -1,22 +1,12 @@
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import React, { Fragment, useState } from 'react';
+import React, { Fragment, useEffect, useState } from 'react';
 
-import Button from 'components/shared/button';
 import Heading from 'components/shared/heading';
 
-import EventsModal from '../events-modal';
+import CardButton from './card-button';
 
 function CardItem({ icon: Icon, name, text, buttons, size }) {
-  const [isOpen, setIsOpen] = useState(false);
-  const openModal = () => {
-    setIsOpen(true);
-  };
-  const closeModal = (e) => {
-    e.stopPropagation();
-    setIsOpen(false);
-  };
-
   const has2Buttons = buttons.length === 2;
   const isSmallSize = size === 'sm';
   return (
@@ -47,19 +37,8 @@ function CardItem({ icon: Icon, name, text, buttons, size }) {
             isSmallSize ? 'xl:gap-x-4' : 'lg:gap-x-5'
           )}
         >
-          {buttons.map(({ buttonUrl, buttonText, buttonTarget, isPopup, calendarId }, index) => (
-            <Fragment key={index}>
-              {isPopup ? (
-                <Button size={size} onClick={openModal}>
-                  {buttonText}
-                </Button>
-              ) : (
-                <Button to={buttonUrl} size={size} target={buttonTarget || ''}>
-                  {buttonText}
-                </Button>
-              )}
-              <EventsModal calendarId={calendarId} isOpen={isOpen} closeModal={closeModal} />
-            </Fragment>
+          {buttons.map((button, index) => (
+            <CardButton {...button} key={index} size={size} />
           ))}
         </div>
       </div>
@@ -74,7 +53,7 @@ CardItem.propTypes = {
   buttons: PropTypes.arrayOf(
     PropTypes.shape({
       buttonText: PropTypes.string.isRequired,
-      buttonUrl: PropTypes.string.isRequired,
+      buttonUrl: PropTypes.string,
       buttonTarget: PropTypes.string,
       isPopup: PropTypes.bool,
       calendarId: PropTypes.string,
